+++ App/models.py
@@ -1,5 +1,6 @@
 from pydantic import BaseModel
 
+
 class UserModel(BaseModel):
     email: str
     password: str
